+++ src/main/java/util/YesOrNo.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import java.util.Arrays;
