+++ src/test/java/util/YesOrNoTest.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import static org.assertj.core.api.Assertions.*;
